pyproject.toml
[project]
name = "linamp"
version = "0.1.0"
description = "Terminal music player - the love child of Winamp and Midnight Commander"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.89.0",
    "python-mpv>=1.0.7",
]

[project.scripts]
linamp = "linamp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
